+++ tests/testthat/test-fit_weight_sim_nereo.R
@@ -1,4 +1,5 @@
 test_that("fit_weight_sim_nereo is a kb_fit_weight the accessors operate on", {
+  expect_s3_class(fit_weight_sim_nereo, "kb_fit_weight_nereo")
   expect_s3_class(fit_weight_sim_nereo, "kb_fit_weight")
   expect_s3_class(fit_weight_sim_nereo, "kb_fit")
   expect_true(nobs(fit_weight_sim_nereo) > 0L)